ssi: aspeed: Automatically configure CS when in user mode
Supermicro's BMC kernel flash driver performs all accesses in user mode,
and seems to expect the chip select be automatically configured. Given
this firmware is supplied on shipping hardware, it appears the hardware
behaves accordingly. Make the model match the expected hardware
behaviour.

The behaviour of asserting the "CE# Stop Active Control" bit on register
update is documented in the datasheet, but undocumented is the automatic
deassert of the "CE# Stop Active Control" bit after configuring user
mode.

With this change the SMC BMC kernel can correctly read the JEDEC ID of
the chip and get as far as attempting to mount root.

Signed-off-by: Andrew Jeffery <andrew@aj.id.au>