hw/misc: Add a iBT device model
Implement an IPMI BT interface model using a chardev backend to
communicate with an external PowerNV guest. The Protocol is the
OpenIPMI simulator protocol for a virtual machine described in :

    https://github.com/cminyard/openipmi/blob/master/lanserv/README.vm

and implemented by the 'ipmi-bmc-extern' model on the host side.

To use, start the Aspeed BMC machine with :

    -chardev socket,id=ipmi0,host=localhost,port=9002,ipv4,server,nowait \
    -global driver=aspeed.ibt,property=chardev,value=ipmi0

and the PowerNV machine with :

    -chardev socket,id=ipmi0,host=localhost,port=9002,reconnect=10 \
    -device ipmi-bmc-extern,id=bmc0,chardev=ipmi0 \
    -device isa-ipmi-bt,bmc=bmc0,irq=10 -nodefaults

This is work in progress. Use with care and kindness, it's probably
quite buggy.

Signed-off-by: Cédric Le Goater <clg@kaod.org>